new file; hitlist.ts
